refactor(comands): move BaseCommand to core
