+++ src/main/kotlin/com/jonaswanke/unicorn/commands/Commands.kt
@@ -1,3 +1,5 @@
 package com.jonaswanke.unicorn.commands
 
+import com.jonaswanke.unicorn.core.BaseCommand
+
 object Commands : BaseCommand()
