+++ src/main/kotlin/com/jonaswanke/unicorn/script/parameters/Argument.kt
@@ -6,7 +6,7 @@ import com.github.ajalt.clikt.core.BadParameterValue
 import com.github.ajalt.clikt.parameters.arguments.*
 import com.github.ajalt.clikt.parameters.options.transformAll
 import com.github.ajalt.clikt.parameters.types.*
-import com.jonaswanke.unicorn.commands.BaseCommand
+import com.jonaswanke.unicorn.core.BaseCommand
 import java.io.File
 import java.nio.file.FileSystem
 import java.nio.file.FileSystems
